# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for SUT SE 2563",
        "title": "SUT SA Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/typediseases": {
            "get": {
                "description": "list typedisease entities",
                "produces": [
                    "application/json"
                ],
                "summary": "List typedisease entities",
                "operationId": "list-typedisease",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.TypeDisease"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "post": {
                "description": "Create typedisease",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create typedisease",
                "operationId": "create-typedisease",
                "parameters": [
                    {
                        "description": "TypeDisease entity",
                        "name": "typedisease",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.TypeDisease"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.TypeDisease"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/typediseases/{id}": {
            "get": {
                "description": "get typedisease by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a typedisease entity by ID",
                "operationId": "get-typedisease",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "TypeDisease ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.TypeDisease"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "put": {
                "description": "update typedisease by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update a typedisease entity by ID",
                "operationId": "update-typedisease",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "TypeDisease ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "TypeDisease entity",
                        "name": "typedisease",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.TypeDisease"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.TypeDisease"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "delete": {
                "description": "get typedisease by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a typedisease entity by ID",
                "operationId": "delete-typedisease",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "TypeDisease ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "ent.Organ": {
            "type": "object",
            "properties": {
                "OrganName": {
                    "description": "OrganName holds the value of the \"OrganName\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the OrganQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.OrganEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.OrganEdges": {
            "type": "object",
            "properties": {
                "forogan": {
                    "description": "Forogan holds the value of the forogan edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Spaciality"
                    }
                }
            }
        },
        "ent.Physician": {
            "type": "object",
            "properties": {
                "PhysicianEmail": {
                    "description": "PhysicianEmail holds the value of the \"PhysicianEmail\" field.",
                    "type": "string"
                },
                "PhysicianName": {
                    "description": "PhysicianName holds the value of the \"PhysicianName\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PhysicianQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PhysicianEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.PhysicianEdges": {
            "type": "object",
            "properties": {
                "foruser": {
                    "description": "Foruser holds the value of the foruser edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Spaciality"
                    }
                }
            }
        },
        "ent.Spaciality": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the SpacialityQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.SpacialityEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.SpacialityEdges": {
            "type": "object",
            "properties": {
                "ogan": {
                    "description": "Ogan holds the value of the ogan edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Organ"
                },
                "type": {
                    "description": "Type holds the value of the type edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.TypeDisease"
                },
                "user": {
                    "description": "User holds the value of the user edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Physician"
                }
            }
        },
        "ent.TypeDisease": {
            "type": "object",
            "properties": {
                "DiseaseName": {
                    "description": "DiseaseName holds the value of the \"DiseaseName\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the TypeDiseaseQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.TypeDiseaseEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.TypeDiseaseEdges": {
            "type": "object",
            "properties": {
                "fortype": {
                    "description": "Fortype holds the value of the fortype edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Spaciality"
                    }
                }
            }
        },
        "gin.H": {
            "type": "object",
            "additionalProperties": true
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information"
            }
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
...
}